changes due to testing the apply button.
User will see the attributes that has been 'ignored' by the hardware when they has been written.
The check to decide if a value has been applied will be repeated some times, until they match or mark it as ignored.
Special check for the floats because differences below the format, the user considers them ok. Some of this differences comes from different float representations in bytes.
the prewrite for near values feature has been changed for integer attributes to try it with +-1.

git-svn-id: svn+ssh://controls01/siciliarep/svn/trunk/tango_client/ctlinac/trunk@13776 a18efa20-c439-0410-bf20-9b91ec695b4c